requirements.txt
Django~=4.2.13
drf-yasg
django-cors-headers
djangorestframework
